refactor: extract checksum, archive and log path builders into a mixin
The eleven methods that build a path from a PdsFile's own path move out of
pdsfile.py into src/pdsfile/_derived_paths.py as `_DerivedPathsMixin`: the
checksum-path group (checksum_path_and_lskip, checksum_path_if_exact,
dirpath_and_prefix_for_checksum), the archive-path group
(archive_path_and_lskip, archive_path_if_exact, dirpath_and_prefix_for_archive,
archive_logpath) and the log-path group (set_log_root, log_path_for_bundle,
log_path_for_bundleset, log_path_for_index).

This commit is a pure move: every one of the eleven definitions is byte-for-byte
what it was, and each of the two contiguous runs compares identical as a single
blob, so nothing was reordered and no blank line was dropped. The mixin holds
methods only and defines no state.

The two source windows are not adjacent. Between them sits the shelf-support
block — SHELF_CACHE, SHELF_ACCESS, SHELF_CACHE_SIZE, SHELF_CACHE_SLOP,
SHELF_ACCESS_COUNT, SHELF_NULL_KEY_VALUES — and it is untouched, as is
LOG_ROOT_, which opens the log-path block. Those are class attributes, which the
Phase-5 preamble keeps on PdsFile, and LOG_ROOT_ and LOGFILE_TIME_FMT are also
frozen manifest members. Both halves of their traffic still work from a mixin
because both go through the class at call time: set_log_root assigns
cls.LOG_ROOT_, and the log_path_for_* methods read cls.LOG_ROOT_ and
cls.LOGFILE_TIME_FMT after cls = type(self).

The free-variable sweep — symtable over each body plus an AST pass over
decorators and argument defaults — names exactly one module-global the moved
code needs, `datetime`, and no module-level constant, function or class. So no
module-level name leaves pdsfile.py, which still uses datetime in its properties
block, and nothing needs a redundant-alias re-export or a deferred import of
PdsFile.

`_DerivedPathsMixin` takes its alphabetical slot in the class statement
(`plans/2026-07-27-addendum-phase5-mixin-base-order.md`), and
tests/api/test_mixin_collisions.py discovers it from PdsFile.__bases__, so its
collision, shadowing, reachability, no-__init__ and base-order checks cover the
new mixin without a new test id.

Ruff ratchet: A002 is the only code the moved lines trigger. All three
occurrences are the frozen `dir=` parameter of the log_path_for_* methods, so
A002 moves wholesale to _derived_paths.py's entry and is removed from
pdsfile.py's. Every other code in pdsfile.py's entry conserves exactly. The
comment over the extracted modules' entries is reworded because the old wording
claimed every code in them also appears in pdsfile.py's entry, which A002 makes
false.
